Move `div_2k_mod_q` ops
